+++ libs/compiler/syntax.h
@@ -51,7 +51,7 @@ typedef struct syntax
 	int startmode;					/**< Start of last record in modetab */
 	
 	int tree[MAXTREESIZE];			/**< Tree */
-	int tc;							/**< Tree counter */
+	size_t tc;							/**< Tree counter */
 	
 	size_t hashtab[256];			/**< Hash table for reprtab */
 	int hash;						/**< Last value of hash function */
